IDFY-128: Alter an old Objective-C class to it's new Swift equivalent.

+++ IDFYListOperationInteractor.swift
@@ -98,7 +98,7 @@ class IDFYListOperationInteractor : IDFYListOperationInteractorInterface {
   
   // MARK: - Convenience functions
   
-  private func indexOfCurrentlyActiveList() -> NSInteger {
+  private func indexOfCurrentlyActiveList() -> Int {
     let listArray = dataManager.getAllLists()
     var indexOfCurrentlyActiveList = -1
     let currentListName = dataManager.getCurrentList().name
